backend/requirements.txt
transformers
datasets
torch
sentence-transformers
hdbscan
umap-learn
streamlit
scikit-learn
snorkel
label-studio-sdk
spacy
scispacy
shap
lime
pandas
